// src/utils/models/faculty-group.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { DepartmentEntry } from './department-entry'
import type { FacultyGroupFacultyType } from './faculty-group-faculty-type'

export interface FacultyGroup {
  facultyName?: string
  facultyType?: FacultyGroupFacultyType
  departments?: DepartmentEntry[]
}